% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfluenceAreaVariable.R
\name{addInfluenceAreaVariable}
\alias{addInfluenceAreaVariable}
\title{function to create a influence area variable from the COD_PUERTO variable}
\usage{
addInfluenceAreaVariable(df, cod_puerto_column)
}
\arguments{
\item{df:}{dataframe to modify}
}
\value{
Return the original dataframe with AREA_INF variable.
}
\description{
Check if COD_PUERTO variable exists and create the AREA_INF variable.
}
