% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arte.R
\docType{data}
\name{arte}
\alias{arte}
\title{List of gears}
\format{A data frame with 2 columns:
\describe{
  \item{COD_ARTE}{Gear code}
  \item{ARTE}{Gear}
  ...
}}
\usage{
arte
}
\description{
A dataset containing the gears used.
}
\keyword{datasets}
