% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupScripts.R
\name{backupScripts}
\alias{backupScripts}
\title{Function make a backup of multiple files to a folder.}
\usage{
backupScripts(files_to_backup, path_files = getwd(), path_backup)
}
\arguments{
\item{files_to_backup}{character vector with files to backup.}

\item{path_files}{path where the files to backup are. By default the actual
working directory.}

\item{path_backup}{path where save the backup files.}
}
\value{
A logical vector of the length of the files_to_backup with TRUE when
the file is properly saved and false if doesn't.
}
\description{
Function make a backup of multiple files to a folder.
}
\details{
If any files of the files_to_backup already exists in path_backup,
the function ask for overwrite all the files.

If the name of the files to backup include a directory, like "data/df1.csv",
the function create the directory first.
}
