% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caladero_origen.R
\docType{data}
\name{caladero_origen}
\alias{caladero_origen}
\title{Fishing ground - Origin}
\format{
A data frame with 3 columns:
\describe{
  \item{CALADERO}{Fishing ground.}
  \item{COD_ORIGEN}{SIRENO origin code.}
  \item{ORIGEN}{Origin description.}
  ...
}
}
\usage{
caladero_origen
}
\description{
A dataset containing the fishing groud with its origin used in SAP MUE team.
}
\keyword{datasets}
