% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchesPercentileUP.R
\name{catchesPercentileUP}
\alias{catchesPercentileUP}
\title{Create dataset with catches percentile for every species.}
\usage{
catchesPercentileUP(dfs, path = getwd())
}
\arguments{
\item{dfs}{vector of filenames with lengths. The file mus be obtained from
'tallas x up' SIRENO reports.}

\item{path}{path where the files are located. Working directory by default.}
}
\value{
dataframe with species, sex, minumum and maximun lengths
}
\description{
This functions works with the catches files obtained by 'Tallas x up' reports
from SIRENO. It's possible user multiple files.
}
\details{
The dataset obtained is used in 'revision volcado' script to check if the catches of
the species are in the range of previous years
}
