% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfpo2015.R
\docType{data}
\name{cfpo2015}
\alias{cfpo2015}
\title{CFPO2015}
\format{A data frame with 32 columns:
\describe{
  \item{DISTINTIVO}{}
  \item{LISTA}{}
  \item{CENSOPORMODALIDAD}{}
  \item{CALADERO}{}
  \item{MODALIDAD}{}
  \item{LABORATORIO_PTOBASE}{}
  \item{CODIGO_PUERTO_BASE}{}
  \item{PUERTO_BASE}{}
  \item{VERNACULO}{}
  \item{PROVINCIA}{}
  \item{AUTONOMIA_PUERTO_BASE}{}
  \item{PAIS_EXPORTACION}{}
  \item{ESTADO}{}
  \item{FECHA_ALTA_CENSO}{}
  \item{FECHA_SITUACION}{}
  \item{FECHA_ENTRADA_SERVICIO}{}
  \item{ARQUEO_TRB}{}
  \item{ARQUEO_GT}{}
  \item{SITUACION_GT}{}
  \item{ESLORA_TOTAL}{}
  \item{ESLORA_PP}{}
  \item{LOA}{}
  \item{POTENCIA_CV}{}
  \item{POTENCIA_KW}{}
  \item{TIPO_CASCO}{}
  \item{IRREGULAR}{}
  \item{AÑO_CONST}{}
  ...
}}
\usage{
cfpo2015
}
\description{
A dataset containing the CFPO2015.
}
\keyword{datasets}
