% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAllFormatVariables.R
\name{checkAllFormatVariables}
\alias{checkAllFormatVariables}
\title{Function to check all the variables format of a dataframe obtained by
importMuestreosUP() function.}
\usage{
checkAllFormatVariables(df)
}
\arguments{
\item{df:}{dataframe with the format variable to check}
}
\value{
TRUE if the format is correct. Return a named list of vectors if there are any error.
The name of the vector is the variable with errors. Every vector contains the
rows index of df with errors.
}
\description{
The format of every variable is in variables_format dataset.
}
