% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/especies.R
\docType{data}
\name{especies}
\alias{especies}
\title{Species}
\format{A data frame with 2 columns:
\describe{
  \item{COD_ESP}{Specie code}
  \item{ESP}{Scientific name}
  ...
}}
\usage{
especies
}
\description{
A dataset with species code and scientific name.
}
\keyword{datasets}
