% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/especies_mezcla.R
\docType{data}
\name{especies_mezcla}
\alias{especies_mezcla}
\title{Mixed species}
\format{
A data frame with 4 columns:
\describe{
  \item{ESP_MUE}{Sampled specie}
  \item{COD_ESP_MUE}{Sampled specie code}
  \item{ESP_CAT}{Category specie}
  \item{COD_ESP_CAT}{Category specie code}
  ...
}
}
\usage{
especies_mezcla
}
\description{
A data set containing the mixed species to be saved in 'Category Species' and
it's corresponding specie in 'Sampled Species'.
}
\keyword{datasets}
