% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/especies_no_mezcla.R
\docType{data}
\name{especies_no_mezcla}
\alias{especies_no_mezcla}
\title{Non mixed species}
\format{A data frame with 2 columns:
\describe{
  \item{ESP}{Specie}
  \item{COD_ESP}{Specie code}
  ...
}}
\usage{
especies_no_mezcla
}
\description{
A dataset containing the non mixed species.
}
\keyword{datasets}
