% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrato_rim.R
\docType{data}
\name{estrato_rim}
\alias{estrato_rim}
\title{List of ESTRATO_RIM available}
\format{A data frame with 3 columns:
\describe{
  \item{ESTRATO_RIM}{}
  \item{VALID}{Control variable}
  ...
}}
\usage{
estrato_rim
}
\description{
A dataset containing the list of ESTRATO_RIM available
}
\keyword{datasets}

