% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrato_rim_arte.R
\docType{data}
\name{estrato_rim_arte}
\alias{estrato_rim_arte}
\title{ESTRATO_RIM - Gear}
\format{
A data frame with 3 columns:
\describe{
  \item{ESTRATO_RIM}{}
  \item{ARTE}{Gear of the ESTRATO_RIM}
  \item{COD_ARTE}{Sireno gear code}
  ...
}
}
\usage{
estrato_rim_arte
}
\description{
A dataset containing the gear of every ESTRATO_RIM used in SAPMUE
}
\keyword{datasets}
