% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estratorim_arte_OAB.R
\docType{data}
\name{estratorim_arte_OAB}
\alias{estratorim_arte_OAB}
\title{ESTRATO_RIM - Gear of OAB samples}
\format{A data frame with 3 columns:
\describe{
  \item{ESTRATO_RIM}{}
  \item{ARTE}{Gear of the ESTRATO_RIM}
  \item{COD_ARTE}{Gear code}
  ...
}}
\usage{
estratorim_arte_OAB
}
\description{
A dataset containing the gears allowed of every ESTRATO_RIM used in OAB
samples.
}
\keyword{datasets}
