% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estratorim_origen.R
\docType{data}
\name{estratorim_origen}
\alias{estratorim_origen}
\title{ESTRATO_RIM - Origin}
\format{
A data frame with 3 columns:
\describe{
  \item{ESTRATO_RIM}{}
  \item{COD_ORIGEN}{SIRENO origin code}
  ...
}
}
\usage{
estratorim_origen
}
\description{
A dataset containing the origins allowed of every ESTRATO_RIM used in SAPMUE
}
\keyword{datasets}
