% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estratorim_origen_OAB.R
\docType{data}
\name{estratorim_origen_OAB}
\alias{estratorim_origen_OAB}
\title{ESTRATO_RIM - Origin of OAB samples}
\format{A data frame with 3 columns:
\describe{
  \item{ESTRATO_RIM}{}
  \item{COD_ORIGEN}{SIRENO origin code}
  ...
}}
\usage{
estratorim_origen_OAB
}
\description{
A dataset containing the origins allowed of every ESTRATO_RIM used in OAB
samples.
}
\keyword{datasets}
