% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCsvSAPMUE.R
\name{exportCsvSAPMUEBASE}
\alias{exportCsvSAPMUEBASE}
\title{export csv with SAPMUE format}
\usage{
exportCsvSAPMUEBASE(df, filename)
}
\arguments{
\item{df:}{dataframe to export}

\item{filename:}{name of the file to export without filename extension.}
}
\description{
The csv files usually used in SAPMUE have a special format:
- fields separation with ";"
- row header
- without quote characteres
- without row names
- decimal separator with ","
}
\details{
This function allow the export with that format.
}

