% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportListToCsv.R
\name{exportListToCsv}
\alias{exportListToCsv}
\title{Export to csv a list of dataframes}
\usage{
exportListToCsv(list, prefix = "", suffix = "", separation = "",
  path = getwd())
}
\arguments{
\item{path}{= destiny path. By default is the actual working directory.}

\item{list:}{list of dataframes}

\item{prefix:}{text to add after the name of the dataframe in the file exported. None by default.}

\item{suffix:}{text to add before the name of the dataframe in the file exported. None by default.}

\item{separation:}{characters to separate the prefix and suffix to the name of
the dataframe. By default is "_".}
}
\value{
Save in the working directory a csv file for each dataframe in the
list of dataframes.
}
\description{
This function create a csv file for each dataframe in a list of dataframes. The
name of every created file contains the name of its corresponding dataframe.
}

