% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportListToGoogleSheet.R
\name{exportListToGoogleSheet}
\alias{exportListToGoogleSheet}
\title{Export a list of dataframes to a google sheet}
\usage{
exportListToGoogleSheet(list, prefix = "", suffix = "",
  separation = "")
}
\arguments{
\item{list:}{list of dataframes}

\item{prefix:}{text to add after the name of the dataframe in the file exported. None by default.}

\item{suffix:}{text to add before the name of the dataframe in the file exported. None by default.}

\item{separation:}{characters to separate the prefix and suffix to the name of
the dataframe. By default is "_".}
}
\value{
Save in the working directory a google sheet file for each dataframe in the
list of dataframes.
The first time this function is run, R console show a link to open in a browser
where a username, password and permissions are asked.
}
\description{
This function create a google sheet file for each dataframe in a list of dataframes and upload it. The
name of every created file contains the name of its corresponding dataframe.
}
