% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportListToXlsx.R
\name{exportListToXlsx}
\alias{exportListToXlsx}
\title{Export to xlsx a list of dataframes}
\usage{
exportListToXlsx(list, prefix = "", suffix = "", separation = "",
  path_export = getwd())
}
\arguments{
\item{list:}{list of dataframes}

\item{prefix:}{text to add after the name of the dataframe in the file exported. None by default.}

\item{suffix:}{text to add before the name of the dataframe in the file exported. None by default.}

\item{separation:}{characters to separate the prefix and suffix to the name of
the dataframe. By default is "_".}

\item{path_export:}{path where the files will be saved.}
}
\value{
Save in the working directory a xlsx file for each dataframe in the
list of dataframes.
}
\description{
This function create a xlsx file for each dataframe in a list of dataframes. The
name of every created file contains the name of its corresponding dataframe.
}
