% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFileAccesoryFunctions.R
\name{formatImportedFile}
\alias{formatImportedFile}
\title{To allow an easiest use of this data in R, variables 'DIA', 'MES', 'YEAR' and
'TRIMESTRE' are added.}
\usage{
formatImportedFile(df)
}
\description{
To allow an easiest use of this data in R, variables 'DIA', 'MES', 'YEAR' and
'TRIMESTRE' are added.
}
