% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFileAccesoryFunctions.R
\name{formatVariableTypes}
\alias{formatVariableTypes}
\title{Format the variables of a dataframe obtained by the RIM and OAB import
funcionts from this package.
The variable types are stored in formato_variables dataset.}
\usage{
formatVariableTypes(df, file_type)
}
\arguments{
\item{df}{dataframe to format.}

\item{file_type}{type of the file: "RIM_CATCHES",
"RIM_CATCHES_IN_LENGTHS", "RIM_LENGTHS", "OAB_TRIPS", "OAB_HAULS",
"OAB_CATCHES" or "OAB_LENGTHS"}
}
\description{
Format the variables of a dataframe obtained by the RIM and OAB import
funcionts from this package.
The variable types are stored in formato_variables dataset.
}
