% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSGPVesselId.R
\name{getSGPVesselId}
\alias{getSGPVesselId}
\title{Get vessel identification code of Spanish SGP (Secretaría General de Pesca)
vessel census.}
\usage{
getSGPVesselId(cfr)
}
\arguments{
\item{cfr}{ship CFR code.}
}
\value{
identification code in SGP census.
}
\description{
Get vessel identification code of Spanish SGP (Secretaría General de Pesca)
vessel census.
}
\note{
The identification code is only used by Spanish SGP.
}
