% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFileAccesoryFunctions.R
\name{getStructureFiles}
\alias{getStructureFiles}
\title{---- Function to get the structure of the files stored in formato_variables dataset
There are two fields related to the type of variables: class_variable_import
and class_variable_final.
The import of the files require the use of the class variables contained in
class_variable_import. Then, some adjustments to certaing variables are made
(replace commas with dots in numeric fiedls...). And finally, the format of the
variables is changed with the class contained in class_variable_final.}
\usage{
getStructureFiles(file, file_type, status_class)
}
\arguments{
\item{file_type}{type of the file to import: "RIM_CATCHES",
"RIM_CATCHES_IN_LENGTHS", "RIM_LENGTHS", "OAB_TRIPS", "OAB_HAULS",
"OAB_CATCHES" or "OAB_LENGTHS"}

\item{status}{"class_variable_import" to get the classes of the variables in
the import proccess of a file; or "class_variable_final" to get final classes
which the variables must have.}
}
\description{
---- Function to get the structure of the files stored in formato_variables dataset
There are two fields related to the type of variables: class_variable_import
and class_variable_final.
The import of the files require the use of the class variables contained in
class_variable_import. Then, some adjustments to certaing variables are made
(replace commas with dots in numeric fiedls...). And finally, the format of the
variables is changed with the class contained in class_variable_final.
}
