% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize.R
\name{humanize}
\alias{humanize}
\title{function to create new description columns from encoded variables of any dataframe
which have any of the encoded variables that is in dataset variables_format.}
\usage{
humanize(df)
}
\arguments{
\item{df:}{dataframe to humanize}
}
\value{
Return the original dataframe with the description columns added.
}
\description{
function to create new description columns from encoded variables of any dataframe
which have any of the encoded variables that is in dataset variables_format.
}
