% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanizeVariable.R
\name{humanizeVariable}
\alias{humanizeVariable}
\title{Humanize variable
Add new column with the description of a variable coded... I mean, of a
encoded column create a new one with the appropriate description.
It's available for variables with a data source: COD_PUERTO, COD_BARCO,
COD_ARTE, COD_ORIGEN, COD_TIPO_MUE, COD_ESP_MUE, COD_ESP_CAT.
Before the humanization, checks the correct format of the variable.}
\usage{
humanizeVariable(df, variable)
}
\arguments{
\item{df:}{df wich contains the variable to humanize}

\item{variable:}{one of this values: COD_PUERTO, COD_BARCO,
COD_ARTE, COD_ORIGEN, COD_TIPO_MUE, COD_ESP_MUE, COD_ESP_CAT}
}
\value{
Return the df dataframe with the new variable added.
}
\description{
Humanize variable
Add new column with the description of a variable coded... I mean, of a
encoded column create a new one with the appropriate description.
It's available for variables with a data source: COD_PUERTO, COD_BARCO,
COD_ARTE, COD_ORIGEN, COD_TIPO_MUE, COD_ESP_MUE, COD_ESP_CAT.
Before the humanization, checks the correct format of the variable.
}
