% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_muestreos_up.R
\name{import_muestreos_up}
\alias{import_muestreos_up}
\title{Import 'muestreos tallas por up'
This function import the 'muestreos tallas por up' files from SIRENO}
\usage{
import_muestreos_up(des_tot, des_tal, tal, by_month = FALSE, export = FALSE)
}
\arguments{
\item{des_tot}{file with the total landings}

\item{des_tal}{file with the landings of the lengths samples}

\item{tal}{file with the lengths samples}

\item{by_month}{to select only one month. MONTH must be fill in the constants section. False by default.}

\item{export}{to export muestreos_up dataframe in csv file. False by default.}
}
\value{
a list with 3 data frames
}
\description{
Import 'muestreos tallas por up'
This function import the 'muestreos tallas por up' files from SIRENO
}

