% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCsvSAPMUE.R
\name{importCsvSAPMUE}
\alias{importCsvSAPMUE}
\title{import csv with usually format used in SAPMUE}
\usage{
importCsvSAPMUE(filename)
}
\arguments{
\item{filename:}{name of the file to be imported.}
}
\value{
A dataframe with the content of the csv, with column names and with
blank fields in rows with unequal length.
}
\description{
The csv files usually used in SAPMUE have a special format:
- fields separation with ";"
- row header
- sometimes the rows have unequal length
}
\details{
This function allow the import with that format.
}
