% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importLengths.R
\name{importLengths}
\alias{importLengths}
\title{Import the Lenghts file from 'tallas por up' report}
\usage{
importLengths(file, path = getwd())
}
\arguments{
\item{file}{vector with the total landings filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the Lengths files obtained from 'muestreos tallas por up'
reports in SIRENO.
}
\details{
Multiple files can be imported at the same time.

To allow a better use of this data in R, fields 'DIA', 'MES', 'YEAR' and 'TRIMESTRE'
has been created.
}
