% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABAccidentals.R
\name{importOABAccidentals}
\alias{importOABAccidentals}
\title{Import SIRENO accidentals catches report from on-board sampling in the ICES
project.}
\usage{
importOABAccidentals(file, path = getwd())
}
\arguments{
\item{file}{vector with the file names.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return data frame.
}
\description{
This function import the accidentals catches file obtained from SIRENO
on-board reports.
}
\details{
Multiple files can be imported at the same time.
}
