% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABCatches.R
\name{importOABCatches}
\alias{importOABCatches}
\title{Import the Catches file from OAB report}
\usage{
importOABCatches(file, path = getwd())
}
\arguments{
\item{file}{vector with the catches filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the catches file obtained from OAB
reports in SIRENO.
}
\details{
Multiple files can be imported at the same time.
}
\note{
The catches SIRENO report must be generated with the option 'by category'
}
