% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABCatchesCECAF.R
\name{importOABCatchesCECAF}
\alias{importOABCatchesCECAF}
\title{Import SIRENO catches report from on-board sampling in the CECAF project.}
\usage{
importOABCatchesCECAF(file, path = getwd())
}
\arguments{
\item{file}{vector with the file names.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return data frame.
}
\description{
This function import the catches file obtained from on-board
reports in SIRENO of CECAF project. The SIRENO catches report must be
generated with the option 'by category'.
}
\details{
Multiple files can be imported at the same time.
}
\note{
The SIRENO catches report must be generated with the option
'by category'.
}
