% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABFilesCECAF.R
\name{importOABFilesCECAF}
\alias{importOABFilesCECAF}
\title{Import the SIRENO on board reports from CECAF project.}
\usage{
importOABFilesCECAF(
  trips,
  hauls,
  catches,
  lengths,
  export = FALSE,
  path = getwd()
)
}
\arguments{
\item{trips}{vector with the names of the haul files.}

\item{hauls}{vector with the names of the trip files.}

\item{catches}{vector with the names of the catch file names.}

\item{lengths}{vector with the names of the length file names.}

\item{export}{to export reports data frame in csv file. False by default.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return a list with 6 data frames.
}
\description{
This function import the on board report files of ICES project from SIRENO:
hauls, trips, catches and lengths files.
}
\note{
The SIRENO catches report must be generated with the option 'by
category'. The SIRENO hauls report must be generated with the
option 'only hauls'. The SIRENO Lengths report must be generated by categories.
}
