% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABLengths.R
\name{importOABLengths}
\alias{importOABLengths}
\title{Import the Lengths file from OAB report}
\usage{
importOABLengths(file, path = getwd())
}
\arguments{
\item{file}{vector with the catches filenames.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the lengths file obtained from OAB
reports in SIRENO. The SIRENO Lengths report must be generated by categories.
}
\details{
Multiple files can be imported at the same time.
}
\note{
The SIRENO Lengths report must be generated by categories.
}
