% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABLengthsCECAF.R
\name{importOABLengthsCECAF}
\alias{importOABLengthsCECAF}
\title{Import SIRENO haul report from on-board sampling in the CECAF project.}
\usage{
importOABLengthsCECAF(file, path = getwd())
}
\arguments{
\item{file}{vector with the file names.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return data frame.
}
\description{
This function import the lengths file obtained from OAB
reports in SIRENO on CECAF project. The SIRENO Lengths report must be
generated by categories.
}
\details{
Multiple files can be imported at the same time.
}
\note{
The SIRENO Lengths report must be generated by categories.
}
