% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRIMCatches.R
\name{importRIMCatches}
\alias{importRIMCatches}
\alias{importCatches}
\title{Import the 'Catches' file from 'tallas por up' report}
\usage{
importRIMCatches(file, path = getwd())
}
\arguments{
\item{file}{vector with the total landings filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the 'Catches' files obtained from 'muestreos tallas por up'
reports in SIRENO.
}
\details{
Multiple files can be imported at the same time.

To allow a better use of this data in R, fields 'DIA', 'MES', 'YEAR' and 'TRIMESTRE'
are created in the returned dataframe.
}
