% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVesselActive.R
\name{isVesselActive}
\alias{isVesselActive}
\title{Indicates if vessel is active in the Spanish SGP (Secretaría General de Pesca)
vessel census.}
\usage{
isVesselActive(cfr)
}
\arguments{
\item{cfr}{ship CFR code.}
}
\value{
True if the ship is active, False if it doesn't.
}
\description{
Indicates if vessel is active in the Spanish SGP (Secretaría General de Pesca)
vessel census.
}
