% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthsRangeUP.R
\name{lengthsRangeUP}
\alias{lengthsRangeUP}
\title{Create dataset with maximun and minimun lengths for every species.}
\usage{
lengthsRangeUP(dfs, path = getwd())
}
\arguments{
\item{dfs}{vector of filenames with lengths. The file mus be obtained from
'tallas x up' SIRENO reports.}

\item{path}{path where the files are located. Working directory by default.}
}
\value{
dataframe with species, sex, minumum and maximun lengths
}
\description{
This functions works with the lenght files obtained by 'Tallas x up' reports
from SIRENO. It's possible user multiple files
}
\details{
The dataset obtained is used in 'revision volcado' script to check if the lengths of
the species are in the range of previous years
}
