% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maestro_categorias.R
\docType{data}
\name{maestro_categorias}
\alias{maestro_categorias}
\title{Categories in SIRENO}
\format{
A data frame with 24016 rows and 8 columns:
\describe{
  \item{ESPCOD}{SIRENO specie code}
  \item{ESPDESTAX}{Specie name}
  \item{PUECOD}{Landing port code}
  \item{PUEDES}{Landing port name}
  \item{ESPCAT}{Category code}
  \item{ESPCATDES}{Category name}
  \item{ESPCATFUE}{}
  \item{TIPPROCOD}{Gutted/whole}
  ...
}
}
\usage{
maestro_categorias
}
\description{
A dataset containing the categories available in SIRENO. The original data
is obtained from SIRENO (report by proyect --> categories)
}
\keyword{datasets}
