% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maestro_flota_sireno.R
\docType{data}
\name{maestro_flota_sireno}
\alias{maestro_flota_sireno}
\title{Fleet in SIRENO}
\format{A data frame with 35,942 rows and 10 columns:
\describe{
  \item{BARCOD}{SIRENO code}
  \item{BARCODSECR}{Secretaría General de Pesca code}
  \item{BARMAT}{Registration Number}
  \item{BARDES}{Vessel name}
  \item{BARFOL}{Registration Number}
  \item{BARFECALTA}{Start date}
  \item{BARFECBAJA}{End date}
  \item{BARESTADO}{Code State}
  \item{PAICOD}{Country code}
  \item{PUECOD}{Port base}
  ...
}}
\usage{
maestro_flota_sireno
}
\description{
A dataset containing the fleet available in SIRENO
}
\keyword{datasets}
