% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metier_caladero_dcf.R
\docType{data}
\name{metier_caladero_dcf}
\alias{metier_caladero_dcf}
\title{Relation between METIER_DCF and CALADERO_DCF}
\format{A data frame with 4 columns:
\describe{
  \item{METIER_DCF}{}
  \item{CALADERO_DCF}{}
  \item{ESTRATO_RIM}{}
  \item{COD_ORIGEN}{}
  ...
}}
\usage{
metier_caladero_dcf
}
\description{
A dataset containing the relation between METIER_DCF and CALADERO_DCF.
}
\keyword{datasets}
