% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metier_coherence.R
\docType{data}
\name{metier_coherence}
\alias{metier_coherence}
\title{Relation between METIER_DCF, CALADERO_DCF, ESTRATO_RIM, ORIGEN and ARTE}
\format{
A data frame with 7 columns:
\describe{
  \item{METIER_DCF}{}
  \item{CALADERO_DCF}{}
  \item{ESTRATO_RIM}{}
  \item{COD_ORIGEN}{}
  \item{ORIGEN}{}
  \item{COD_ARTE}{}
  \item{ARTE}{}
  ...
}
}
\usage{
metier_coherence
}
\description{
A data set containing the relation between METIER_DCF, CALADERO_DCF, ORIGEN
and ARTE.
}
\keyword{datasets}
