% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metiers.R
\docType{data}
\name{metiers}
\alias{metiers}
\title{Data set of metiers. Include the old ones (2009-2020) and its relation with
the new ones.}
\format{
A data frame with columns:
\describe{
  \item{METIER_DCF}{}
  \item{METIER_DCF_2009_2020}{}
  \item{COMMENTS}{}
  ...
}
}
\usage{
metiers
}
\description{
Data set of metiers. Include the old ones (2009-2020) and its relation with
the new ones.
}
\keyword{datasets}
