% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origen.R
\docType{data}
\name{origen}
\alias{origen}
\title{Fishing origin of the samples}
\format{A data frame with 2 columns:
\describe{
  \item{COD_ORIGEN}{code used in SIRENO}
  \item{ORIGEN}{name of the origin}
  ...
}}
\usage{
origen
}
\description{
A dataset containing the fishing origin of the samples.
}
\keyword{datasets}

