% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p90_percentil_historico.R
\docType{data}
\name{p90_capturas_historico}
\alias{p90_capturas_historico}
\title{Historical percentile 90 range by species.}
\format{A data frame with 4 columns:
\describe{
  \item{COD_ESP}{SIRENO code species}
  \item{SEXO}{sex}
  \item{TALLA_MIN}{minimun length}
  \item{TALLA_MAX}{maximun length}
  ...
}}
\usage{
p90_capturas_historico
}
\description{
A dataset containing the historical percentile 90 range by species. The dataset is created
in create_p90_capturas_historico.R, using catches recorded in SIRENO between 2009
to 2016.
}
\keyword{datasets}
