% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p97_percentil_historico.R
\docType{data}
\name{p97_capturas_historico}
\alias{p97_capturas_historico}
\title{Historical percentile 97 range by species.}
\format{
A data frame with 4 columns:
\describe{
  \item{COD_ESP}{SIRENO code species}
  \item{ESTRATO_RIM}{RIM stratum}
  \item{P97}{percentile 97}
  ...
}
}
\usage{
p97_capturas_historico
}
\description{
A dataset containing the historical percentile 97 of species catcges. The dataset is created
in create_p97_capturas_historico.R, using catches recorded in SIRENO between 2014
to 2018.
}
\keyword{datasets}
