% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p99_percentil_historico.R
\docType{data}
\name{p99_capturas_historico}
\alias{p99_capturas_historico}
\title{Historical percentile 99 range by species.}
\format{A data frame with 4 columns:
\describe{
  \item{COD_ESP}{SIRENO code species}
  \item{ESTRATO_RIM}{RIM stratum}
  \item{P99}{percentile 99}
  ...
}}
\usage{
p99_capturas_historico
}
\description{
A dataset containing the historical percentile 99 of species catcges. The dataset is created
in create_p99_capturas_historico.R, using catches recorded in SIRENO between 2014
to 2018.
}
\keyword{datasets}
