% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescripciones_oab_2021_coherencia.R
\docType{data}
\name{prescripciones_oab_2021_coherencia}
\alias{prescripciones_oab_2021_coherencia}
\title{Prescriptions coherence.}
\format{
A data frame with 9 columns.
}
\usage{
prescripciones_oab_2021_coherencia
}
\description{
A dataset containing the coherence between the variables of 2021 OAB
prescriptions. The original prescriptions document contain grouped variables
like ports ports like "Aguas Comunitarias no españolas".
}
\details{
In this dataframe, those groups has been disagregated.

And other variables has been added in order to get a complete dataset where
find the allowed coherence between them in prescriptions.
}
\keyword{datasets}
