% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescripciones_rim_mt2_2021_coherencia.R
\docType{data}
\name{prescripciones_rim_mt2_2021_coherencia}
\alias{prescripciones_rim_mt2_2021_coherencia}
\title{Prescriptions coherence.}
\format{
A data frame with 11 columns.
}
\usage{
prescripciones_rim_mt2_2021_coherencia
}
\description{
A dataset containing the coherence between the variables of 2021 RIM
prescriptions. Tho original prescriptions document contain grouped variables
like fisheries with multiple rim stratum o grouped ports like "Avilés/Gijón".
}
\details{
In this dataframe, those groups has been disagregated.

And other variables has been added in order to get a complete dataset where
find the coherence between them allowed in prescriptions.
}
\keyword{datasets}
