% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedencia.R
\docType{data}
\name{procedencia}
\alias{procedencia}
\title{Sampling institution}
\format{
A data frame with 1 rows and 2 columns:
\describe{
  \item{PROCEDENCIA}{sampling institution}
  \item{VALID}{control variable}
  ...
}
}
\usage{
procedencia
}
\description{
A dataset containing the sampling institution.
}
\keyword{datasets}
