% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puerto.R
\docType{data}
\name{puerto}
\alias{puerto}
\title{Port}
\format{
A data frame with 25 rows and 2 columns:
\describe{
  \item{PUERTO}{port name}
  \item{COD_PUERTO}{code use in SIRENO}
  \item{LOCODE}{LOCODE code}
  ...
}
}
\usage{
puerto
}
\description{
A dataset containing ports used.
}
\keyword{datasets}
