% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rango_tallas_historico.R
\docType{data}
\name{rango_tallas_historico}
\alias{rango_tallas_historico}
\title{Historical species size range by sex.}
\format{A data frame with 4 columns:
\describe{
  \item{COD_ESP}{SIRENO code species}
  \item{SEXO}{sex}
  \item{TALLA_MIN}{minimun length}
  \item{TALLA_MAX}{maximun length}
  ...
}}
\usage{
rango_tallas_historico
}
\description{
A dataset containing the historical size range by sex. The dataset is created
in create_sizeRangeHistorical.R, using lengths recorded in SIRENO between 2009
to 2016.
}
\keyword{datasets}
