% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rango_tallas_historico_OAB.R
\docType{data}
\name{rango_tallas_historico_OAB}
\alias{rango_tallas_historico_OAB}
\title{Historical size range by species in OAB}
\format{
A data frame with 4 columns:
\describe{
  \item{COD_ESP}{SIRENO code species}
  \item{SEXO}{sex}
  \item{TALLA_MIN}{minimun length}
  \item{TALLA_MAX}{maximun length}
  ...
}
}
\usage{
rango_tallas_historico_OAB
}
\description{
A dataset containing the historical size range by species and sex of OAB samples.
The dataset is created
in create_rango_tallas_historico.R, using lengths recorded in SIRENO between 2014
and 2018.
}
\keyword{datasets}
