% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relacion_variables.R
\docType{data}
\name{relacion_variables}
\alias{relacion_variables}
\title{To improve legibility, some field names from the reports obtained in SIRENO are
changed in the import process.
This dataset contain the relation betwen the fields before and after the import process}
\format{
A data frame with 9 columns:
\describe{
  \item{original_name_variable}{name of the variable before the import process}
  \item{name_variable}{name of the variable after the import process}
  ...
}
}
\usage{
relacion_variables
}
\description{
To improve legibility, some field names from the reports obtained in SIRENO are
changed in the import process.
This dataset contain the relation betwen the fields before and after the import process
}
\keyword{datasets}
