% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSppFromSp.R
\name{removeSppFrom_Sp}
\alias{removeSppFrom_Sp}
\title{Remove " spp" abbreviation in species.}
\usage{
removeSppFrom_Sp(sp)
}
\arguments{
\item{sp}{A character vector of species.}
}
\value{
A character vector of species without " spp".
}
\description{
Remove " spp" abbreviation in species.
}
\author{
Marco A. Ámez , \email{marco.amez@ieo.es}
}
