% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spBelongsToTaxon.R
\name{spBelongsToTaxon}
\alias{spBelongsToTaxon}
\title{Check in WORMS if one species belongs to a taxon.}
\usage{
spBelongsToTaxon(species, taxon_to_match)
}
\arguments{
\item{taxon_to_match}{Taxon to match.}

\item{specie}{Specie to check.}
}
\value{
TRUE if the species belong to taxon. False if doesn't. In case any species
doesn't exists in WORMS a warning is thrown and the text "this species
doesn't match in WORMS" is returned. In case multiple matchs, a warning is
thrown and the text "multiple match in WORMS" is returned.
}
\description{
Check in WORMS if one species belongs to a taxon.
}
\details{
Uses WORMS api to identify the species taxon.

If the species contains the abbreviation spp, the " spp" is removed in
order to find a match in WORMS.

Require jsonlite library.
}
\author{
Marco A. Ámez , \email{marco.amez@ieo.es}
}
