% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo_muestreo.R
\docType{data}
\name{tipo_muestreo}
\alias{tipo_muestreo}
\title{Type of sample}
\format{
A data frame with 2 rows and 2 columns:
\describe{
  \item{COD_TIPO_MUE}{code use in SIRENO}
  \item{TIPO_MUE}{description}
  \item{ESTRATEGIA}{sample strategy}
  ...
}
}
\usage{
tipo_muestreo
}
\description{
A dataset containing type of samples used.
}
\keyword{datasets}
