% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_to_humanize.R
\docType{data}
\name{variables_to_humanize}
\alias{variables_to_humanize}
\title{Variables to humanize}
\format{A data frame with 4 columns:
\describe{
  \item{ORIGINAL_VAR}{Variable suitable to humanize.}
  \item{MASTER_VAR}{Name of the ORIGINAL_VAR in the master. This column is
  neccesary because in some cases the name of the variable in the master
  is different that the ORIGINAL_VAR.}
  \item{HUMANIZED_VAR}{Name of the humanized variable. The humanized variable
  is the value returned by humanized functions.}
  \item{MASTER}{Master in wich the variable is coded.}
  ...
}}
\usage{
variables_to_humanize
}
\description{
A dataset with the variables suitable to humanize.
Contains to the data necessary to allow the humanize process.
}
\keyword{datasets}

