% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vesselInSGPFleet.R
\name{vesselInSGPFleet}
\alias{vesselInSGPFleet}
\title{Detect if a vessel exists in Spanish SGP (Secretaría General de Pesca)
census.}
\usage{
vesselInSGPFleet(cfr)
}
\arguments{
\item{cfr}{ship CFR code.}
}
\value{
True if the vessel exists, False if it doesn't.
}
\description{
Detect if a vessel exists in Spanish SGP (Secretaría General de Pesca)
census.
}
