% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas_influencia.R
\docType{data}
\name{areas_influencia}
\alias{areas_influencia}
\title{List of influence areas}
\format{
A data frame with 4 columns:
\describe{
  \item{PUERTO}{}
  \item{COD_PUERTO}{Port code}
  \item{AREA_INF}{Influence area}
  ...
}
}
\usage{
areas_influencia
}
\description{
A data set containing the list of influence areas.
}
\keyword{datasets}
