% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAllFormatVariables.R
\name{checkAllFormatVariables}
\alias{checkAllFormatVariables}
\title{Function to check all the variables format of a dataframe.}
\usage{
checkAllFormatVariables(df)
}
\arguments{
\item{df:}{dataframe with the format variable to check}
}
\value{
TRUE if the format is correct. Return a named list of vectors if there are any mistake.
The name of the vector is the variable with errors. Every vector contains list of index with errors.
}
\description{
The right format of every variable is in variables_format dataset.
}

