% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFormatVariable.R
\name{checkFormatVariable}
\alias{checkFormatVariable}
\title{Function to check the correct format of a variable of a dataframe
obtained by importMuestreosUP() function.}
\usage{
checkFormatVariable(df, var_to_check)
}
\arguments{
\item{df:}{dataframe with the format variable to check}

\item{var_to_check:}{variable to check}
}
\value{
TRUE if the format is correct. Return a named list of vectors if there are any mistake.
The name of the vector is the variable with errors. Every vector contains list of index with errors.
}
\description{
The format of the variables is in dataset variables_format.
}
