% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSNtoLogical.R
\name{convertSNtoLogical}
\alias{convertSNtoLogical}
\title{Convert logical varible in spanish (S/N) to logical TRUE/FALSE.}
\usage{
convertSNtoLogical(x)
}
\arguments{
\item{x}{vector to convert. Only admit "S", "N" or empty values. Empty values
are assumed as FALSE.}
}
\value{
vector var converted as logical. In case a value is distinct of "S",
"N" or empty value, an error is thrown.
}
\description{
Convert logical varible in spanish (S/N) to logical TRUE/FALSE.
}
