% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/especies_sexadas.R
\docType{data}
\name{especies_sexadas}
\alias{especies_sexadas}
\title{Sexed species}
\format{A data frame with 4 columns:
\describe{
  \item{COD_ESP}{Sampled specie sireno code}
  \item{ESP_MUE}{Sampled specie name}
  \item{COD_PUERTO}{Sampled port sireno code}
  \item{PUERTO}{Sampled port name}
  \item{LOCODE}{Sampled port LOCODE code}
}}
\usage{
especies_sexadas
}
\description{
A dataset containing the species which are sexed in the sampling.
}
\keyword{datasets}

