% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportListToCsv.R
\name{exportListToCsv}
\alias{exportListToCsv}
\title{Export to csv a list of dataframes}
\usage{
exportListToCsv(list)
}
\arguments{
\item{list}{of dataframes}
}
\value{
Save in the working directory a csv file for every dataframe in the
list of dataframes.
}
\description{
This function create a csv file every dataframe in a list of dataframes.
}

