% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixReportSirenoFiles.R
\name{fixReportSirenoFiles}
\alias{fixReportSirenoFiles}
\title{Fix erroneus format of the reports from SIRENO in the 'observations' variable.
Sometimes in SIRENO, in 'observations' variable, a semicolon or a enter is
saved. When SIRENO generate the report files, the 'observations' variable is
exported as it is, with the semicolon and enter characteres. This means that
some lines of the file have more than the theorical number of variables (when
observations variable have a semicolon inside) or less than the theorical
(when a enter character is in the observations variable of the previous row).}
\usage{
fixReportSirenoFiles(filename, filetype, path = getwd(), temporal = F)
}
\arguments{
\item{filename}{Filename of the file to check.}

\item{filetype}{Tipe of file: RIM_CATCHES, RIM_CATCHES_IN_LENGTHS,
RIM_LENGTHS,OAB_TRIPS, ,OAB_HAULS, OAB_CATCHES, OAB_LENGTHS, OAB_LITTER,
OAB_ACCIDENTAL}

\item{path}{Path of the file.}

\item{temporal}{If is True, the file is exported in a temporal directory. This is
usefull when this function is used with sapmbuenase import functions. When is
false, the fixed report is exported in the path location}
}
\description{
This function check the number of variables and works as follow:
- If the number of variables is the same of the theorical, there aren't any
changes.
- When there are more variables than theorical, the last variables are
collapsed in the last one and changed the ";" to "." inside the observations
variable
- When there are less variable than the theorical, we asume an enter has
been saved in the observations variable of the previous row and after
the enter there are more semicolons. The content of this row is added to the
variable observation of the previous row and the ";" is changed to ".".
}
\note{
Use this function carefully, if the format or the sireno reports
changed can produce unexpected results. Furthermore, only is appropiate when
the 'observations' field is the last one of the row.
}
