% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFileAccesoryFunctions.R
\name{getVariableTypes}
\alias{getVariableTypes}
\title{Get the variable types of the RIM and OAB files downloaded from SIRENO.
This function use the formato_variables dataset.
Used in the import and format of the files.}
\usage{
getVariableTypes(file_type, status_class)
}
\arguments{
\item{file_type}{type of the file: "RIM_CATCHES",
"RIM_CATCHES_IN_LENGTHS", "RIM_LENGTHS", "OAB_TRIPS", "OAB_HAULS",
"OAB_CATCHES" or "OAB_LENGTHS"}

\item{status_class}{"class_variable_import" to get the classes of the variables in
the import proccess of a file; or "class_variable_final" to get final classes
which the variables must have.}
}
\value{
dataframe with the name of the variable, status class and file type.
}
\description{
Get the variable types of the RIM and OAB files downloaded from SIRENO.
This function use the formato_variables dataset.
Used in the import and format of the files.
}
