% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize.R
\name{humanize}
\alias{humanize}
\title{function to create new variables with description values from coded variables.}
\usage{
humanize(df)
}
\arguments{
\item{df:}{dataframe to humanize}
}
\value{
Return the original dataframe with description variables.
}
\description{
Check if any of this variables exists in the dataframe:
- COD_PUERTO
- COD_BARCO
- COD_ARTE --> not available yet
- COD_ORIGEN --> not available yet
- COD_TIPO_MUE --> not available yet
- COD_ESP_MUE
- COD_CATEGORIA --> not available yet
- COD_ESP_CAT --> not available yet
}
\details{
If it exists, create a new variable with its corresponding description values
}

