% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIPDFile.R
\name{importIPDFile}
\alias{importIPDFile}
\title{Import IPD file}
\usage{
importIPDFile(filename, by_month = FALSE)
}
\arguments{
\item{filename}{name of the IPD file}

\item{by_month}{to filter only by one month. If false nothing is filtered. False by default.}
}
\value{
Return a data frame.
}
\description{
This function import the data of biological samples keyed by IPD
}
