% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importIPDFile.R
\name{importIPDFile}
\alias{importIPDFile}
\title{Import IPD file}
\usage{
importIPDFile(filename, by_month = FALSE, path = getwd())
}
\arguments{
\item{filename}{name of the IPD file}

\item{by_month}{to filter only by one month. If false nothing is filtered. False by default.}

\item{path}{= destiny path. By default is the actual working directory.}
}
\value{
Return data frame.
}
\description{
This function import the data of biological samples saved by IPD. Only import
record of the samples of ICES proyect.
}

