% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMuestreosUP.R
\name{importMuestreosUP}
\alias{importMuestreosUP}
\title{Import 'muestreos tallas por up'}
\usage{
importMuestreosUP(des_tot, des_tal, tal, by_month = FALSE, export = FALSE,
  path = getwd())
}
\arguments{
\item{des_tot}{file with the total landings}

\item{des_tal}{file with the landings of the lengths samples}

\item{tal}{file with the lengths samples}

\item{by_month}{to filter only by one month. Numeric between 1 to 12 to select
one month or FALSE for all the year. FALSE by default.}

\item{export}{to export muestreos_up dataframe in csv file. False by default.}
}
\value{
Return a list with 3 data frames
}
\description{
This function import the three files obtained from 'muestreos tallas por up' reports in SIRENO.
}
