% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABAccidentals.R
\name{importOABAccidentals}
\alias{importOABAccidentals}
\title{Import the Litter file from OAB report}
\usage{
importOABAccidentals(file, path = getwd())
}
\arguments{
\item{file}{vector with the Accidentals filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the Accidentals file obtained from SIRENO OAB reports.
}
\details{
Multiple files can be imported at the same time.
}
