% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABFiles.R
\name{importOABFiles}
\alias{importOABFiles}
\title{Import OAB reports from SIRENO.}
\usage{
importOABFiles(mareas, lances, capturas, tallas, export = FALSE,
  path = getwd())
}
\arguments{
\item{mareas}{vector with hauls filenames}

\item{lances}{vector with the trips filenames}

\item{capturas}{vector with the catches filenames}

\item{tallas}{vectir with the lengths filenames}

\item{export}{to export muestreos_up dataframe in csv file. False by default.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return a list with 3 data frames
}
\description{
This function import the four OAB report files from SIRENO: hauls file,
trips file, catches files and lengths file.
}
\note{
The SIRENO catches report must be generated with the option 'by category'
}
