% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABHauls.R
\name{importOABHauls}
\alias{importOABHauls}
\title{Import the Hauls file from OAB report}
\usage{
importOABHauls(file, path = getwd())
}
\arguments{
\item{file}{vector with the hauls filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the hauls file obtained from OAB
reports in SIRENO.
}
\details{
Multiple files can be imported at the same time.
}
