% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABLitter.R
\name{importOABLitter}
\alias{importOABLitter}
\title{Import the Litter file from OAB report}
\usage{
importOABLitter(file, path = getwd())
}
\arguments{
\item{file}{vector with the catches filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the litter file obtained from OAB
reports in SIRENO.
}
\details{
Multiple files can be imported at the same time.
}
\note{
The SIRENO catches report must be generated with the option 'by category'
}
