% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABTrips.R
\name{importOABTrips}
\alias{importOABTrips}
\title{Import the Trips file from OAB report}
\usage{
importOABTrips(file, path = getwd())
}
\arguments{
\item{file}{vector with the trips filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the trips file obtained from OAB
reports in SIRENO.
}
\details{
Multiple files can be imported at the same time.
}
