% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importOABTripsCECAF.R
\name{importOABTripsCECAF}
\alias{importOABTripsCECAF}
\title{Import the CECAF Trips file from OAB report}
\usage{
importOABTripsCECAF(file, path = getwd())
}
\arguments{
\item{file}{vector with the trips filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return dataframe.
}
\description{
This function import the trips file obtained from OAB
reports in SIRENO from CECAF project.
}
\details{
Multiple files can be imported at the same time.
}
