% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRIMFiles.R
\name{importRIMFiles}
\alias{importRIMFiles}
\alias{importMuestreosUP}
\title{Import 'muestreos tallas por up'}
\usage{
importRIMFiles(des_tot, des_tal, tal, by_month = FALSE, export = FALSE,
  path = getwd())
}
\arguments{
\item{des_tot}{vector with the total landings filenames}

\item{des_tal}{vector with the landings of the lengths samples filenames}

\item{tal}{vector with the lengths samples filenames}

\item{by_month}{to filter only by one month. Numeric between 1 to 12 to select
one month or FALSE for all the year. FALSE by default.}

\item{export}{to export muestreos_up dataframe in csv file. False by default.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return a list with 3 data frames
}
\description{
This function import the three files obtained from 'muestreos tallas por up'
reports in SIRENO: total catches file, catches of length samples file and lengths
file.
}
\details{
To allow a better use of this data in R, fields 'DIA', 'MES', 'YEAR' and 'TRIMESTRE'
are created in the returned dataframe.
}
