% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRIMFiles.R
\name{importRIMFiles}
\alias{importRIMFiles}
\alias{importMuestreosUP}
\title{Import 'muestreos tallas por up'}
\usage{
importRIMFiles(catches, catches_in_lengths, lengths, by_month = FALSE,
  export = FALSE, path = getwd())
}
\arguments{
\item{catches}{vector with the total landings filenames}

\item{catches_in_lengths}{vector with the landings of the lengths samples filenames}

\item{lengths}{vector with the lengths samples filenames}

\item{by_month}{to filter only by one month. Numeric between 1 to 12 to select
one month or FALSE for all the year. FALSE by default.}

\item{export}{to export muestreos_up dataframe in csv file. False by default.}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return a list with 3 data frames
}
\description{
This function import the three files obtained from 'muestreos tallas por up'
reports in SIRENO: total catches file, catches of length samples file and lengths
file.
}
