% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formato_variables.R, R/puerto.R
\docType{data}
\name{puerto}
\alias{puerto}
\title{Variables format}
\format{A data frame with 51 rows and 9 columns:
\describe{
  \item{name_variable}{name of the variable}
  \item{regex_variable_import}{regular expression of the variable in the import process}
  \item{class_variable_import}{class of the variable in the import process}
  \item{mandatory}{TRUE if the variable is mandatory for the import process}
  \item{regex_variable_final}{regular expression that should be the variable after import process}
  \item{class_variable_final}{class that should be the variable after import process}
  \item{CATCHES}{TRUE if the variable is in catches dataframe}
  \item{CATCHES_IN_LENGTHS}{TRUE if the variable is in catches_in_lengths dataframe}
  \item{LENGTHS}{TRUE if the variable is in lengths dataframe}
  ...
}}
\usage{
puerto

puerto
}
\description{
A dataset containing the format of the variables in the tallas_x_up files

A dataset containing ports used.
}
\keyword{datasets}

