% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesBelongToTaxon.R
\name{speciesBelongToTaxon}
\alias{speciesBelongToTaxon}
\title{Return species belonging to certain taxon.}
\usage{
speciesBelongToTaxon(species, taxon)
}
\arguments{
\item{species}{vector of species to check.}

\item{taxon}{taxon.}
}
\value{
vector with the species belonging to taxon. For every species, TRUE
if belongs to taxon. False if doesn't. In case the species
doesn't exists in WORMS a warning is thrown and the text "this species
doesn't match in WORMS" is returned. In case multiple matchs, a warning is
thrown and the text "multiple match in WORMS" is returned.
}
\description{
Return species belonging to certain taxon.
}
\details{
Is the vectorized version of check_spe_belongs_to_taxon().
Uses WORMS api to identify the species taxon.
If the species contains the abbreviation spp, the " spp" is removed in
order to find a match in WORMS.
}
\author{
Marco A. Ámez , \email{marco.amez@ieo.es}
}
